tests/wpt/tests/html/canvas/offscreen/path-objects/2d.path.fill.overlap.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.path.fill.overlap</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.path.fill.overlap</h1>
<p class="desc"></p>


<script>
var t = async_test("");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.fillStyle = '#000';
  ctx.fillRect(0, 0, 100, 50);

  ctx.fillStyle = 'rgba(0, 255, 0, 0.5)';
  ctx.rect(0, 0, 100, 50);
  ctx.closePath();
  ctx.rect(10, 10, 80, 30);
  ctx.fill();

  _assertPixelApprox(canvas, 50,25, 0,127,0,255, 1);
  t.done();

});
</script>
